botelle/botelle_backend/requirements.txt
gunicorn==20.0.4
Flask==1.1.2
Flask-API==2.0
Flask-PyMongo==2.3.0
flask-mongoengine==0.9.5
bleach==3.1.5
dnspython==1.16.0
shortuuid==1.0.1
Flask-RESTful==0.3.8
PyJWT==1.7.1
py-bcrypt==0.4
Flask-JWT-Extended==3.24.1
sendgrid==6.3.1
google-cloud-secret-manager==1.0.0
isort==4.3.21
firebase-admin==4.3.0
py-healthcheck==1.10.0